% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_area.R
\name{db_area}
\alias{db_area}
\title{Du Bois area plot}
\usage{
db_area(
  data,
  order,
  cat1,
  cat2,
  dpi = 320,
  seed = 42,
  res_step = 0.001,
  limits = c(-1, 1),
  names = c(cat1, cat2),
  title,
  subtitle,
  message,
  path = getwd(),
  filename
)
}
\arguments{
\item{data}{\strong{data-frame class object:} the data to be shown in the
area plot. Each row must contain the value of each of the two categories
as well one instance of the order variable.}

\item{order}{\strong{character:} the name of the order variable.}

\item{cat1}{\strong{character:} the name of the leftmost category (in red).
Its values are highlighted in the vertical axis.}

\item{cat2}{\strong{character:} the name of the rightmost category (in black).}

\item{dpi}{\strong{integer:} resolution of the rendered texts and image.}

\item{seed}{\strong{integer:} seed used for calculate the bounded random walk.}

\item{res_step}{\strong{numeric:} the product of \code{res_step} and the
numeric range of the order variable defines the step size of the bounded
random walk.}

\item{limits}{\strong{numeric vector:} the lower and superior limits that
bound the random walk}

\item{names}{\strong{character vector:} the names of the categories. shown in
the first areas in the superior part of the plot and in the vertical axis
title in the left.}

\item{title}{\strong{character:} text on the top of the plot}

\item{subtitle}{\strong{character:} text under the title}

\item{message}{\strong{character:} text on the bottom of the plot. has a
brown background.}

\item{path}{\strong{character:} path to save the image of the plot.}

\item{filename}{\strong{character:} name of the image and its extension.}
}
\value{
an 22x28 inches image of an area plot in Du Bois style
}
\description{
Creates an area plot with two categories (\code{cat1} and \code{cat2}) over
the values of an \code{order} variable. They are shown in terms of percentage
and their sum is 100\%. The values of the categories are associated with the
horizontal axis. Their areas are stacked horizontally. The rightmost category
(in red) has its values highlighted in the vertical axis on the right. The
order variable is shown in the vertical axis and is labeled in the left. Both
axes are inverted. A bounded random walk effect can be added to the left of
the area (like a the plot has been torn) as function of some parameters
(\code{seed}, \code{res_step} and \code{limits}). Furthermore, there is three
slots for text (\code{title}, \code{subtitle} and \code{message})
}
\examples{
library(dplyr)
library(tidyr)
library(magrittr)
data <- managers \%>\%
  dplyr::select(race, year, pct_bosses_total) \%>\%
  tidyr::pivot_wider(names_from = "race",
  values_from = "pct_bosses_total")

db_area(data = data, order = "year", cat1 = "black", cat2 = "white",
  limits = c(-3,4), filename = "managers.png",
  title = "PARTICIPATION IN MANAGERIAL POSITIONS BY RACE IN BRAZIL.",
  subtitle = "INSPIRED BY: W.E.B. DU BOIS | DATA FROM: IBGE | GRAPHIC BY: ICARO BERNARDES",
  message = "IN THE SERIES, USUALLY WHITES OCCUPY SLIGHTLY LESS GENERAL WORK POSITIONS. HOWEVER WHITES OCCUPY WAY MORE MANAGERIAL POSITIONS THAN BLACKS")
}
