% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{managers}
\alias{managers}
\title{Blacks and whites participation in the total number of managers and common
workers in Brazil.}
\format{
A data frame with 14 rows and 7 variables:
\describe{
\item{total_bosses_per_1000}{total number of workers (all races) that are managers per 1000 people}
\item{total_workers_per_1000}{total number of workers (all races, managers or not) per 1000 people}
\item{race}{race of the workers}
\item{year}{year of the survey}
\item{pct_bosses_total}{percentage of workers from a particular race that are managers}
\item{pct_workers_total}{percentage of workers (managers or not) from a particular race}
\item{cv_pct_bosses}{coefficient of variance of the percentage of managers}
}
}
\source{
\url{https://ftp.ibge.gov.br/Indicadores_Sociais/Desigualdades_por_Cor_ou_Raca/xls/2_Trabalho_xls.zip} - Tabela 2.7.xls
}
\usage{
managers
}
\description{
Percentage of people in the workforce in general and managerial positions by
race. Data from Brazil collected by IBGE (Brazilian Govt. Statistics Agency).
}
\keyword{datasets}
